Roll AI contract renewals per matchday instead of in one season-start batch
The bulk-at-season-start processor worked but the all-or-nothing
timing felt unnatural — every AI renewal landed on Aug 15 and the
pre-contract pool was binary: either locked-in by Jan 1 or wide open
for the whole window. A per-matchday roll mirrors how real clubs keep
extending their stars throughout the year and drains the pool
gradually so the user encounters fewer pre-contract opportunities as
the window progresses.

- New RollAIContractRenewals listener subscribed to GameDateAdvanced,
  alongside the existing RollSalaryUnhappiness / ApplyWageGapMoraleDrip
  per-tick listeners.
- Per-tick rates calibrated to ~75% / ~45% / ~10% cumulative renewal
  by Jan 1 (~20 ticks in) for top / mid / low importance — matching
  the end-state the bulk processor produced, but spread over time.
- Skips veterans, loaned-out players, and players who already have a
  pending or agreed pre-contract from another club so the AI can't
  rug-pull a user mid-negotiation.
- Raw DB read-pass scoped to teams that actually have an eligible
  expiring player — most ticks across most clubs short-circuit at
  the distinct-team-ids query (~5–10ms) and do no further work.

Removes AIPreemptiveRenewalProcessor (and its test) — the per-tick
approach replaces it outright; running both would double-renew.
