+++ app/Modules/Season/Services/SeasonSetupPipeline.php
@@ -4,7 +4,6 @@ namespace App\Modules\Season\Services;
 
 use App\Modules\Season\Contracts\SeasonProcessor;
 use App\Modules\Season\DTOs\SeasonTransitionData;
-use App\Modules\Season\Processors\AIPreemptiveRenewalProcessor;
 use App\Modules\Season\Processors\BudgetProjectionProcessor;
 use App\Modules\Season\Processors\ContinentalAndCupInitProcessor;
 use App\Modules\Season\Processors\LeagueFixtureProcessor;
@@ -34,7 +33,6 @@ class SeasonSetupPipeline
         YouthAcademyPromotionProcessor $academyPromotion,
         LeagueFixtureProcessor $fixtureGeneration,
         StandingsResetProcessor $standingsReset,
-        AIPreemptiveRenewalProcessor $aiPreemptiveRenewal,
         BudgetProjectionProcessor $budgetProjection,
         ContinentalAndCupInitProcessor $competitionInitialization,
         UefaSuperCupQualificationProcessor $uefaSuperCup,
@@ -48,7 +46,6 @@ class SeasonSetupPipeline
             $academyPromotion,
             $fixtureGeneration,
             $standingsReset,
-            $aiPreemptiveRenewal,
             $budgetProjection,
             $competitionInitialization,
             $uefaSuperCup,
